added energy transport; addressed some PR comments
Alongside gwf and gwt, we now have gwe in this model. This is mostly done as a proof of concept and does not lend much to the paired visualizations with the tabletop model.

Docstrings and comments have been added liberally.

New modpath architecture for animating the flow of the groundwater model

New plotting infrastructure that lets the user decide which dependent variable they would like so see plotted in the animation via the "parameter" input of make_the_animation()