Layered groundwater model/mpk (#1)
* update to .gitignore to include the .idea folder

* Added bin/exes, functioning model framework, too

main.py constructs and runs the models specified by the sheets of the scenarios.xlsx file.

Next steps:

make this model specific to the construction/layout of the tabletop model.

flesh out the scenarios

add the transport models

construct the animations

* house keeping, functionality unchanged

* Functional flow, path, and transport models

There's still plenty of refinement in the solution, visualization, and scenarios, but all three models are functional!

* model refinements, plotting updates

* Create environment.yml

Added the base environment file that we use for our EStL models. This provides a foundation environment.yml file that we can modify as needed.

* first iteration of polygon plotting, added riv cells

* editing the yml

aimed at removing the unnecessary geospatial things and adding back some of the more basic packages we might need.

* new workflow to not use geopandas

* gridlines updated

* river illustration, well 9, scenarios_short3.xlsx

the river channel is now included in the topology for coloration/visualization.

well 9 is now included in the scenarios files

* housekeeping, add gif for readme

also, set up for scenario 1, a long, flow only simulation for continuous display/looping and scenario 2, a shorter demo of contaminant transport from well 2 to well 6 in the confined aquifer.

* Update README.md

add gif to readme

* swap .mp4 in instead of .gif for saving

* added energy transport; addressed some PR comments

Alongside gwf and gwt, we now have gwe in this model. This is mostly done as a proof of concept and does not lend much to the paired visualizations with the tabletop model.

Docstrings and comments have been added liberally.

New modpath architecture for animating the flow of the groundwater model

New plotting infrastructure that lets the user decide which dependent variable they would like so see plotted in the animation via the "parameter" input of make_the_animation()

* up the resolution of the output animation

* addressing review comments

AZ recommended simplifying the input schedule files --> scenarios.xlsx

AEJ noted extra folders were being created outside of the "outputs" folder --> fixed

AEJ noted that we likely won't want any outputs folders tracked so a change was made to the .gitignore to reflect.

---------

Co-authored-by: allanejones <alljones@illinois.edu>